apps/api/requirements.txt
# Core FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Vector database
pgvector==0.2.4

# Redis for caching and queues
redis==5.0.1
celery==5.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML dependencies
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langgraph==0.0.20
openai==1.3.7
anthropic==0.7.8

# Smart contract analysis tools
slither-analyzer==0.9.3
semgrep==1.50.0

# WebSocket support
websockets==12.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
ruff==0.1.6
mypy==1.7.1
pre-commit==3.6.0

# Monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
prometheus-client==0.19.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# JSON and data processing
orjson==3.9.10
pandas==2.1.4

# Blockchain and Ethereum
web3==6.11.3
eth-account==0.9.0
eth-utils==2.2.0
